shutdown executor after the runner finishes
If the runner instantiates its own executor, then it is also
responsible for calling 'shutdown'. This was the cause of a
bug in the notebook integration where kernels would fail to die
if ProcessPoolExecutors were not properly shut down. This is probably
a bug in IPython, but we fix our usage here also.