add 'shutdown_executor' parameter to runner
Now any executor used by the runner is, by default, shutdown when
the runner exits. The most common usage of passing a custom executor
seems to be to limit the number of cores, and in any case people
generally create 1 executor per runner. Adding 'shutdown_excutor'
as a parameter still allows people with more advanced usecases
to still manage the executor themselves.