remove ipyparallel and distributed as dependencies
We provide compatibility to these libraries, but they are not
dependencies. If the user just wants to use all the cores on their
machine, 'concurrent.futures' is good enough. If they want to use
a cluster, they will have to learn to set up ipyparallel or distributed
themselves *anyway*, so we don't save users any effort by adding the
packages as dependencies