ensure atomic writes when saving a file
Right now, if a program crashes in the middle of saving, you lose
all your data.

Right now, if a program crashes in the middle of saving, you lose
all your data. This ensures that the old file is first moved, then
the new file is saved, and only then the old file is removed.