Fix NaN issue for Learner1D R -> R^n
If a function evaluates over a point where it is not defined, NaN will be returned.
In the case of a function R -> R^n, it caused `_update_scale` to compute `self._scale[1] = nan`,
which in turns led to the evaluation of the function on all other points to be NaN, causing an
infinite loop, as reported in issue #339.

Using `np.nanmin` and `np.nanmax` in place of `np.min` and `np.max` seems to solve the issue.